RedThicketReadMe
Jill Plansky

Overview & Use Details: This is a business website for Red Thicket Rabbitry. The home page has a carousel of  bunny pictures. The home page also showcases the breeds raised, if you hover over the pictures with the mouse the breed name for the rabbits appears. The navigation bar is located in the middle of the page. The website includes a Contact Us form submission for people to submit questions and comments. Through formspree, the Contact Us submission is sent my mailbox.  Social media links are located at the top of the page, when you hover the icon the color changes. If you click on the link it takes you to the rabbitry social media sites.

Technologies used: HTML, CSS, Javascript, and Bootstrap

Future Improvement: Listed also on drawio. These enhancements will fulfill the needs from the user stories. 
1. Rabbit Care Section
2. Rabbit Nutrition
3. Instagram Image Gallery

More enhancements not listed in drawio
4. A link listed on each image of rabbits to adopt. This link would link to show a documents, which is a pedigree
5. Ecommerce section with affiliate links to Amazon and Chewy. Also a section with products made by Red Thicket Rabbitry